+++ src/cobib/parsers/doi.py
@@ -69,19 +69,20 @@ class DOIParser(Parser):
         doi = match.group(1)
         LOGGER.info("Gathering BibTex data for DOI: %s.", doi)
         try:
-            page = requests.get(DOI_URL + doi, headers=DOI_HEADER, timeout=10)
+            session = requests.Session()
+            page = session.get(DOI_URL + doi, headers=DOI_HEADER, timeout=10)
             if page.encoding is None:
                 page.encoding = "utf-8"
             # this assumes that the doi.org page redirects to the correct journal's landing page
             redirected_url: str = ""
-            header = requests.head(DOI_URL + doi, timeout=1).headers
+            header = session.head(DOI_URL + doi, timeout=1).headers
             LOGGER.debug("The DOI URL header: '%s'", header)
             max_iter = 3
             while "Location" in header and max_iter:
                 max_iter -= 1
                 redirected_url = header["Location"]
                 LOGGER.debug("The found URL redirects to: '%s'", redirected_url)
-                header = requests.head(redirected_url, timeout=1).headers
+                header = session.head(redirected_url, timeout=1).headers
         except requests.exceptions.RequestException as err:
             LOGGER.error("An Exception occurred while trying to query the DOI: %s.", doi)
             LOGGER.error(err)