+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - a hint in the documentation how to filter for (non-)existence of fields (#141)
 - the coBib command `prompt` to the TUI's command palette (!310)
 
+### Changed
+- HTTP requests are now done inside of `requests.Session` instances (!233)
+
 ### Fixed
 - the `Escape` key during an input prompt triggered after a `Confirm` prompt (e.g. a cancelled `quit`)
 - the behavior of the `Home` and `End` keybindings in the TUI's search view (!310)